requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
coverage==7.6.1
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.8
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
orjson==3.10.15
packaging==25.0
pandas==2.0.3
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.19.2
pytest==8.3.5
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.2.1
typer==0.16.1
typing-extensions==4.13.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
